Add post to index functionality complete
This is fairly self explanatory. Up to 15 posts (modifiable through a
global variable) are now added to the index file and when the number
of posts equals 15 on the index page, the oldest page gets pushed out.
Next feature set to be completed is to add to the archive after which
the main set of functionality will be done