Add link to archive.html
Once the month and date are checked for and inserted into the archive
file, the script then inserts a link to the post into the archives.
The final step here is to create the archive file itself and add posts
to it which honestly, I think might be overkill. I don't like the
design so I might skip it for now and keep it just in develop and move
everything else to master.