Finalized Base Generator Code
This commit finalizes the basic code that does the following steps
each time a post is published:

Add a timestamp to the post.

Add the text to the index: In the future this needs to be truncated
text. Not the whole thing. Although again this is kinda debatable. I
really want to give people a painless reading experience but at the
same time if I want to encourage them to comment on the articles then
clicking through would be the best thing. Someday I'll figure out how
to integrate A/B testing into this thing and then we'll know for sure.
Till then it stays as is

Check if the time of the post exists in the archive.html file. This
process actually has an issue around the way it is coded which I
realized a little later. If a post gets pushed for publishing at
something like 23:59:59.90 on the last day of a month/year (new year
resolution posts anyone?) it'll most likely get captured as 2014
January while the post itself might be 2013 December. But this seems
like such an edge case I thought I'd handle it later while refactoring

Add the year and month to the archive file if it doesn't exist and
then add any posts under that created month.

And that's it. Still a lot of features to develop from here but from
this point it's pretty simple to start publishing and I will be doing
just that