Added Disqus thread Comment Count
The feature to show "Posted on <date> <comment count> was supposed to
be a feature in the original package as well but I actually missed
this when implmenting the software. This commit adds support for it.

It also might bring about the closing stage for the initial framework
after which this branch can probably be deleted. One thing that has
arisen after the intial framework was added was a lot of code being
duplicated and therefore there's some need for refactoring which I
didn't forsee when putting down the pseudocode.

On the todo list from this point,
Refactor the code to make use of more common methods to
- Get the post time
- Inser the post time
- Write the files

There is also a pressing concern that the code may need to be
rewritten a bit since I'd like to bring in the prettify method when
writing to a file. This will be tested in a later commit though