Added basic RSS 2.0 Feed Creation
Used the module PyRSS2Gen for this because I didn't think it was
necessary to recreate the entire process for myself when such a stable
and easy module exists. The RSS feed being created right now however
is extremely basic and gives just the most essential information
needed. Will continue to update this branch as time goes by